# docs/api-docs.json
{"openapi":"3.0.0","info":{"title":"LightsOn API","version":"1.0.0","description":"Documentazione API per LightsOn"},"servers":[{"url":"http://localhost:5000","description":"Server di sviluppo"}],"components":{"securitySchemes":{"cookieAuth":{"type":"apiKey","in":"cookie","name":"lo_access_token"}}},"paths":{"/me":{"get":{"summary":"Restituisce il ruolo dell'utente autenticato","tags":["Auth"],"security":[{"cookieAuth":[]}],"responses":{"200":{"description":"Dati utente","content":{"application/json":{"schema":{"type":"object","properties":{"userId":{"type":"string"},"email":{"type":"string","format":"email"},"profile":{"type":"object","properties":{"firstName":{"type":"string"},"lastName":{"type":"string"}}},"role":{"type":"string"}}}}}},"401":{"description":"Non autenticato","content":{"application/json":{"schema":{"type":"object","properties":{"error":{"type":"string"}}}}}},"404":{"description":"Utente non trovato","content":{"application/json":{"schema":{"type":"object","properties":{"msg":{"type":"string"}}}}}},"500":{"description":"Errore del server","content":{"application/json":{"schema":{"type":"object","properties":{"error":{"type":"string"}}}}}}}}},"/logout":{"post":{"summary":"Logout (rimuove il cookie di accesso)","tags":["Auth"],"responses":{"200":{"description":"Logout effettuato con successo","content":{"application/json":{"schema":{"type":"object","properties":{"msg":{"type":"string"}}}}}},"500":{"description":"Errore del server","content":{"application/json":{"schema":{"type":"object","properties":{"error":{"type":"string"}}}}}}}}},"/light/{lightId}":{"get":{"summary":"Recupera la segnalazione attiva di un lampione","tags":["Reports"],"parameters":[{"in":"path","name":"lightId","required":true,"schema":{"type":"string"},"description":"ID del lampione"}],"responses":{"200":{"description":"Dati della segnalazione attiva (o null se non presente)","content":{"application/json":{"schema":{"type":"object","properties":{"activeReport":{"oneOf":[{"type":"object","properties":{"reportId":{"type":"string"},"approvedCounts":{"type":"integer"},"description":{"type":"string"},"state":{"type":"string"}}},{"type":"null"}]}}}}}},"404":{"description":"Lampione non trovato","content":{"application/json":{"schema":{"type":"object","properties":{"message":{"type":"string"},"activeReport":{"type":"null"}}}}}},"500":{"description":"Errore del server","content":{"application/json":{"schema":{"type":"object","properties":{"message":{"type":"string"},"activeReport":{"type":"null"}}}}}}}}},"/approve":{"post":{"summary":"Approva (vota) una segnalazione","tags":["Reports"],"security":[{"cookieAuth":[]}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","required":["reportId"],"properties":{"reportId":{"type":"string","description":"ID della segnalazione da approvare"}}}}}},"responses":{"200":{"description":"Voto aggiunto con successo","content":{"application/json":{"schema":{"type":"object","properties":{"message":{"type":"string"},"approvedCounts":{"type":"integer"}}}}}},"400":{"description":"reportId mancante o utente ha già votato","content":{"application/json":{"schema":{"type":"object","properties":{"message":{"type":"string"},"alreadyVoted":{"type":"boolean"}}}}}},"404":{"description":"Report non trovato","content":{"application/json":{"schema":{"type":"object","properties":{"message":{"type":"string"}}}}}},"500":{"description":"Errore del server","content":{"application/json":{"schema":{"type":"object","properties":{"message":{"type":"string"}}}}}}}}},"/new_report":{"post":{"summary":"Crea una nuova segnalazione per un lampione","tags":["Reports"],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","required":["lightId","userId","description","malfunctionType"],"properties":{"lightId":{"type":"string"},"userId":{"type":"string"},"description":{"type":"string"},"malfunctionType":{"type":"string"}}}}}},"responses":{"201":{"description":"Report creato con successo","content":{"application/json":{"schema":{"type":"object","properties":{"message":{"type":"string"},"reportId":{"type":"string"}}}}}},"500":{"description":"Errore del server","content":{"application/json":{"schema":{"type":"object","properties":{"message":{"type":"string"}}}}}}}}},"/all":{"get":{"summary":"Recupera tutte le segnalazioni (admin)","tags":["Admin"],"responses":{"200":{"description":"Lista di tutte le segnalazioni formattate","content":{"application/json":{"schema":{"type":"object","properties":{"reports":{"type":"array","items":{"type":"object","properties":{"id":{"type":"string"},"lightId":{"type":"string"},"date":{"type":"string","format":"date-time"},"problemType":{"type":"string"},"status":{"type":"string"},"upvoteCount":{"type":"integer"},"description":{"type":"string"},"producer":{"type":"object","properties":{"name":{"type":"string"},"phone":{"type":"string"}}}}}}}}}}},"500":{"description":"Errore del server","content":{"application/json":{"schema":{"type":"object","properties":{"message":{"type":"string"}}}}}}}}},"/status":{"post":{"summary":"Aggiorna lo stato di una segnalazione (admin)","tags":["Admin"],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","required":["reportId","status"],"properties":{"reportId":{"type":"string"},"status":{"type":"string","enum":["pending","working on","resolved"]}}}}}},"responses":{"200":{"description":"Stato aggiornato con successo","content":{"application/json":{"schema":{"type":"object","properties":{"message":{"type":"string"},"status":{"type":"string"}}}}}},"400":{"description":"Valore di stato non valido","content":{"application/json":{"schema":{"type":"object","properties":{"message":{"type":"string"}}}}}},"404":{"description":"Report non trovato","content":{"application/json":{"schema":{"type":"object","properties":{"message":{"type":"string"}}}}}},"500":{"description":"Errore del server","content":{"application/json":{"schema":{"type":"object","properties":{"message":{"type":"string"}}}}}}}}},"/lights":{"get":{"summary":"Recupera i lampioni in base alla viewport e al livello di zoom (step)","tags":["Lights"],"parameters":[{"in":"query","name":"minLat","required":true,"schema":{"type":"number"},"description":"Latitudine minima del bounding box"},{"in":"query","name":"maxLat","required":true,"schema":{"type":"number"},"description":"Latitudine massima del bounding box"},{"in":"query","name":"minLng","required":true,"schema":{"type":"number"},"description":"Longitudine minima del bounding box"},{"in":"query","name":"maxLng","required":true,"schema":{"type":"number"},"description":"Longitudine massima del bounding box"},{"in":"query","name":"step","schema":{"type":"string","default":"1"},"description":"Livello di zoom (1, 2 o 3). Determina l'area effettiva e la fascia esterna da considerare."},{"in":"query","name":"viewId","schema":{"type":"string"},"description":"Identificativo della vista (utile per il frontend)"}],"responses":{"200":{"description":"Lista di lampioni nell'area richiesta","content":{"application/json":{"schema":{"type":"object","properties":{"viewId":{"type":"string","description":"Identificativo della vista passato come parametro"},"step":{"type":"integer","description":"Livello di zoom convertito in intero"},"lights":{"type":"array","items":{"type":"object","properties":{"id":{"type":"string","description":"ID del lampione"},"position":{"type":"array","items":{"type":"number"},"description":"Coordinate [latitudine, longitudine]"},"approvedCounts":{"type":"integer","description":"Numero di segnalazioni approvate per questo lampione"}}}}}}}}},"500":{"description":"Errore del server","content":{"application/json":{"schema":{"type":"object","properties":{"message":{"type":"string"}}}}}}}}},"/lights/age":{"get":{"summary":"Calcola l'età di ogni lampione in base alla data di installazione","tags":["Lights"],"responses":{"200":{"description":"Lista di lampioni con età calcolata, ordinati per età decrescente","content":{"application/json":{"schema":{"type":"array","items":{"type":"object","properties":{"id":{"type":"string","description":"ID del lampione"},"manufacturer":{"type":"string","description":"Produttore del lampione"},"installationDate":{"type":"string","format":"date","description":"Data di installazione"},"age":{"type":"number","nullable":true,"description":"Età in anni (approssimata a un decimale) o null se data mancante"},"coordinates":{"type":"array","items":{"type":"number"},"description":"Coordinate geografiche [longitudine, latitudine]"}}}}}}},"500":{"description":"Errore del server","content":{"application/json":{"schema":{"type":"object","properties":{"error":{"type":"string"}}}}}}}}},"/lights/stats":{"get":{"summary":"Statistiche globali dei lampioni (attivi, inattivi, interventi in corso)","tags":["Lights"],"responses":{"200":{"description":"Oggetto con statistiche e interventi","content":{"application/json":{"schema":{"type":"object","properties":{"total":{"type":"integer","description":"Numero totale di lampioni"},"active":{"type":"integer","description":"Numero di lampioni attivi (senza segnalazioni attive)"},"inactive":{"type":"integer","description":"Numero di lampioni inattivi (con segnalazione attiva)"},"interventions":{"type":"array","description":"Interventi in corso o in attesa","items":{"type":"object","properties":{"lightId":{"type":"string","description":"ID del lampione"},"status":{"type":"string","enum":["pending","working on"],"description":"Stato della segnalazione"},"description":{"type":"string","description":"Descrizione della segnalazione"},"reportId":{"type":"string","description":"ID della segnalazione"}}}}}}}}},"500":{"description":"Errore del server","content":{"application/json":{"schema":{"type":"object","properties":{"error":{"type":"string"}}}}}}}}},"/signup":{"post":{"summary":"Registrazione nuovo utente","tags":["Auth"],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","required":["email","password","firstName","lastName"],"properties":{"email":{"type":"string","format":"email"},"password":{"type":"string","minLength":8},"firstName":{"type":"string"},"lastName":{"type":"string"}}}}}},"responses":{"201":{"description":"Utente creato con successo, token JWT nel cookie","content":{"application/json":{"schema":{"type":"object","properties":{"msg":{"type":"string"},"role":{"type":"string"}}}}}},"400":{"description":"Campi mancanti o email già esistente","content":{"application/json":{"schema":{"type":"object","properties":{"msg":{"type":"string"}}}}}},"500":{"description":"Errore del server","content":{"application/json":{"schema":{"type":"object","properties":{"error":{"type":"string"}}}}}}}}},"/login":{"post":{"summary":"Login utente","tags":["Auth"],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","required":["email","password"],"properties":{"email":{"type":"string","format":"email"},"password":{"type":"string"}}}}}},"responses":{"200":{"description":"Login riuscito, token JWT nel cookie","content":{"application/json":{"schema":{"type":"object","properties":{"msg":{"type":"string"},"role":{"type":"string"}}}}}},"400":{"description":"Credenziali errate","content":{"application/json":{"schema":{"type":"object","properties":{"msg":{"type":"string"}}}}}},"500":{"description":"Errore del server","content":{"application/json":{"schema":{"type":"object","properties":{"error":{"type":"string"}}}}}}}}}},"tags":[]}